[CSF-10] Routing añadido
