+++ src/styles/CandidatesDashboard.module.css
